requirements.txt
uvicorn == 0.21.0
pytest == 7.2.1
requests == 2.28.2
fastapi == 0.95.1

pydantic~=1.10.6
starlette~=0.26.1
psycopg2-binary==2.9.5
httpx==0.24.0
confluent-kafka==2.1.0